Broke out signal_last to helper.py and added unit tests
